Introduce :ignore option to has_many :dependent option. If :dependent is set to :ignore,   dependent objects will not be deleted or nullified but just ignored.

+++ lib/simply_stored/instance_methods.rb
@@ -152,7 +152,10 @@ module SimplyStored
             case property.options[:dependent]
             when :destroy
               dependent.destroy
-            else
+            when :ignore
+              # skip
+            else 
+              # nullify
               unless dependent.class.soft_deleting_enabled?
                 dependent.send("#{self.class.foreign_property}=", nil)
                 dependent.save(false)
